+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## 1.1 (work in progress)
 
+- [x] Libraries
+- [x] Search books
 - [x] Improve title formatting
+- [x] Search on Goodreads
 
 
 ## 1.0